+++ mutations/fu_T1_03_counts_towards_progress_uses_viewer.pl
@@ -1,0 +1,1 @@
+s/if \(self::counts_towards_progress\(\$cm, \$ownerid\)\) \{/if (self::counts_towards_progress(\$cm, \$viewerid)) {/;