+++ mutations/fu_T1_09_own_card_skips_is_locked.pl
@@ -1,0 +1,1 @@
+s/        if \(\$ownerid === \$viewerid\) \{\n            return self::is_locked\(\$course, \$viewerid\);\n        \}\n\n//;