+++ mutations/fu_T1_10_lockdate_uses_owner.pl
@@ -1,0 +1,1 @@
+s/\$availabilitydate = self::get_availability_date\(\$course, \$viewerid\);/\$availabilitydate = self::get_availability_date(\$course, \$ownerid);/;