+++ mutations/fu_T1_11_enrolstart_uses_owner.pl
@@ -1,0 +1,1 @@
+s/\$enrolstartdate = self::get_enrolment_start_date\(\$course, \$viewerid\);/\$enrolstartdate = self::get_enrolment_start_date(\$course, \$ownerid);/;