+++ mutations/fu_T1_14_owner_argument_ignored.pl
@@ -1,0 +1,1 @@
+s/\$ownerid = \$userid > 0 \? \$userid : \$viewerid;/\$ownerid = \$viewerid;/;