+++ mutations/fu_T1_15_progress_service_drops_owner.pl
@@ -1,0 +1,1 @@
+s/\$data = static::progress_data\(\(int\) \$courseid, \$ownerid\);/\$data = static::progress_data((int) \$courseid, 0);/;