+++ mutations/fu_T1_18_status_iscompleted_uses_viewer.pl
@@ -1,0 +1,1 @@
+s/is_course_complete\(\$ownerid\)/is_course_complete(\$USER->id)/;