+++ mutations/fu_T1_21_read_plan_dropped.pl
@@ -1,0 +1,1 @@
+s/\$plan = self::read_plan\(\$planid\);\n        \$ownerid/\$plan = new plan(\$planid);\n        \$ownerid/;