+++ mutations/fu_T1_22_can_read_user_dropped.pl
@@ -1,0 +1,1 @@
+s/if \(!user_competency::can_read_user\(\$ownerid\)\) \{/if (false) {/;