+++ mutations/fu_T1_23_scope_check_dropped.pl
@@ -1,0 +1,1 @@
+s/\n        self::require_competency_in_scope\(\$plan, \$competencyid\);\n/\n/;