+++ mutations/fu_T1_34_js_progress_drops_plan.pl
@@ -1,0 +1,1 @@
+s/args: \{courseids: \[courseid\], planid: planid, competencyid: competencyid\}/args: {courseids: [courseid]}/;