+++ mutations/fu_T1_35_js_status_drops_plan.pl
@@ -1,0 +1,1 @@
+s/args: \{courseids: courseIds, planid: planid, competencyid: competencyid\}/args: {courseids: courseIds}/;